refactor(users): move shift constants to validations module
- Relocate DEFAULT_SHIFT_SCHEDULE and TIMEZONE_OPTIONS out of the
  "use client" editor into lib/validations/user.ts so they can be
  imported from Server Components without pulling in client code.
- Switch the editor's per-row error lookup from indexOf(day) to the
  findIndex() that already located the row, so error mapping stays
  correct if day ordering ever changes.
