+++ app/(dashboard)/dashboard/users/_components/user-form.tsx
@@ -12,12 +12,13 @@ import { parseDateInputValue, toDateInputValue } from "@/lib/date-input";
 import {
 	type CreateUserInput,
 	createUserSchema,
+	DEFAULT_SHIFT_SCHEDULE,
 	type ShiftScheduleFieldErrors,
 	type ShiftScheduleInput,
 	type UpdateUserInput,
 	updateUserSchema,
 } from "@/lib/validations/user";
-import { DEFAULT_SHIFT_SCHEDULE, ShiftScheduleEditor } from "./shift-schedule-editor";
+import { ShiftScheduleEditor } from "./shift-schedule-editor";
 
 const inputClass =
 	"block w-full rounded-xl border border-gray-200 dark:border-white/10 bg-gray-50/50 dark:bg-white/5 px-4 py-3 text-sm text-foreground placeholder:text-muted-foreground focus:outline-none focus:bg-card focus:ring-4 focus:ring-primary/30 focus:border-primary transition-all duration-200";
